requirements.txt
gspread>=5.0.0
google-auth>=2.0.0
python-docx>=1.0.0
streamlit>=1.35.0
docx2pdf>=0.1.8
pypdf>=3.0.0
